completed work on the search, apply filter and clear filter
