+++ public/index.html
@@ -11,6 +11,10 @@
     <h1>Joe's Super Awesome Website</h1>
     <p>This is a super duper website. ONLY 5 Ratings <3</p>
     <div id="app"></div> 
+    <script type="text/babel" src="./components/Modale.js"></script> 
+    <script type="text/babel" src="./components/Searches.js"></script> 
+    <script type="text/babel" src="./components/Table.js"></script> 
+    <script type="text/babel" src="./components/TableRow.js"></script> 
     <script type="text/babel" src="./app.js"></script> 
   </body> 
 </html>
